Project router and controller created
